refactor(typescript): Migrate get-project-content to typescript
